pyproject.toml
[project]
name = "test-shiny-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic[bedrock]>=0.42.0",
    "chatlas>=0.2.0",
    "jsonref>=1.1.0",
    "mcp>=1.1.2",
    "openai>=1.58.0",
    "pyyaml>=6.0.2",
    "shiny>=1.2.1",
    "starlette>=0.42.0",
    "uvicorn>=0.34.0",
]

[tool.uv.sources]
chatlas = { git = "https://github.com/posit-dev/chatlas" }
